Clean up announcements and resources
